# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payment Gateway API",
        "description": "API untuk payment gateway simple dengan integrasi Midtrans. \r\n *     \r\n *     Features:\r\n *     - JWT Authentication\r\n *     - Shopping Cart Management\r\n *     - Order Processing\r\n *     - Payment via Midtrans\r\n *     - Order Reminder (Scheduled)",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        },
        {
            "url": "https://api.example.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/orders/checkout": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Checkout - Buat order dari cart",
                "description": "Proses checkout: validasi stok, buat order, generate payment token Midtrans, kurangi stok produk",
                "operationId": "fe9b55a49d82c434f029698c42c13c80",
                "responses": {
                    "200": {
                        "description": "Checkout berhasil - Order created dan payment token generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Checkout successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "order_number": {
                                                    "type": "string",
                                                    "example": "ORD-1234567890"
                                                },
                                                "total_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 250000
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "snap_token": {
                                                    "description": "Token untuk Midtrans Snap payment",
                                                    "type": "string",
                                                    "example": "abc123def456"
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "product_id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "product_name": {
                                                                "type": "string",
                                                                "example": "Product Name"
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 100000
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - Cart kosong atau stok tidak cukup",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Cart is empty."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/history": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Lihat riwayat order",
                "description": "Get semua order user dengan pagination",
                "operationId": "93cc88af1cfedcd8d848c13dd6f427cf",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah item per page (default 15, max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success - Return paginated order history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "History successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 10
                                                            },
                                                            "order_number": {
                                                                "type": "string",
                                                                "example": "ORD-1234567890"
                                                            },
                                                            "total_amount": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 250000
                                                            },
                                                            "status": {
                                                                "description": "pending, paid, failed, expired",
                                                                "type": "string",
                                                                "example": "pending"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-03-20T10:30:00Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/midtrans/webhook": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Webhook dari Midtrans",
                "description": "Receive payment notification dari Midtrans, diproses async via queue 'webhooks' (high priority)",
                "operationId": "403159a05459ba4fa7d8b504eafeffcb",
                "requestBody": {
                    "description": "Payload dari Midtrans payment notification",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "status_code",
                                    "gross_amount",
                                    "signature_key",
                                    "transaction_status"
                                ],
                                "properties": {
                                    "order_id": {
                                        "description": "Order number",
                                        "type": "string",
                                        "example": "ORD-1234567890"
                                    },
                                    "status_code": {
                                        "description": "Status code dari Midtrans",
                                        "type": "string",
                                        "example": "200"
                                    },
                                    "gross_amount": {
                                        "description": "Total amount",
                                        "type": "string",
                                        "example": "250000.00"
                                    },
                                    "signature_key": {
                                        "description": "Signature untuk validasi",
                                        "type": "string",
                                        "example": "abc123..."
                                    },
                                    "transaction_status": {
                                        "description": "capture, settlement, pending, deny, expire, cancel",
                                        "type": "string",
                                        "example": "settlement"
                                    },
                                    "fraud_status": {
                                        "description": "Optional: accept, challenge, deny",
                                        "type": "string",
                                        "example": "accept"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook diterima dan queued for processing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Webhook received and queued for processing"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error - Payload tidak lengkap"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}